move multi tenancy to lib
