+++ .env.example
@@ -1,0 +1,5 @@
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=apps
+DB_USER=user
+DB_PASSWORD=password
