--- apps/identity/src/config/index.ts
@@ -1,3 +1,0 @@
-export * from "./definitions";
-export * from "./database-builder";
-export * from "./multi-schema";
