+++ libs/multi-tenancy/src/definitions/index.ts
