+++ libs/multi-tenancy/src/definitions/interface.ts
