+++ tsconfig.json
@@ -18,11 +18,11 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@app/database": [
-        "libs/database/src"
+      "@app/multi-tenancy": [
+        "libs/multi-tenancy/src"
       ],
-      "@app/database/*": [
-        "libs/database/src/*"
+      "@app/multi-tenancy/*": [
+        "libs/multi-tenancy/src/*"
       ]
     }
   }
